Don't manage namespace if namespace is set to kube-system
This is a safe-guard against the component grabbing ownership of
namespace `kube-system`, if this namespace is configured somewhere in
the hierarchy.

Users may want to install the csi-cloudscale plugin in `kube-system`, as
pods with priority class `system-cluster-critical` can only be created
in `kube-system` in K8s prior to 1.17. For 1.17+ the
`system-cluster-critical` priority class needs to be explicitly enabled
for namespaces other than kube-system using an admission controller
config file, cf. [1].

[1] https://kubernetes.io/docs/concepts/policy/resource-quotas/#limit-priority-class-consumption-by-default