Configure resource requests and limits
The component now sets the default resource requests and limits for all
containers of the CSI driver and controller.

By default, the component sets requests of 20 milli-CPU and 32 MiB, and
limits of 100 milli-CPU and 128 MiB for each container.

The commit also adds a parameter which allows users to adjust the
requests and limits for individual containers. Note that the component
doesn't validate the provided values, and it's the user's responsibility
to provide values which result in valid container specs.

Fixes #24